% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_outcome.R
\name{extract.results}
\alias{extract.results}
\title{Compare and compute the results for a single run.}
\usage{
extract.results(true.model, estimated.model)
}
\description{
Compare and compute the results for a single run.
}
