% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{get.architecture}
\alias{get.architecture}
\title{Generate an undirected unweighted graph.}
\usage{
get.architecture(type, nodes, positive.edge.ratio, ...)
}
\description{
Generate an undirected unweighted graph.
}
