% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising_empirical.R
\name{ising_empirical}
\alias{ising_empirical}
\title{Use Ising parameters from an empirical network architecture.}
\usage{
ising_empirical()
}
\description{
This function returns the Ising model parameters for an empirical network architecture.
}
\details{
The parameters are from: https://doi.org/10.1371/journal.pone.0167490
}
