% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising_random.R
\name{ising_random}
\alias{ising_random}
\title{Generate Ising parameters for a random network architecture.}
\usage{
ising_random(number_nodes, p, seed = F)
}
\arguments{
\item{number_nodes}{(int scalar) Number of nodes.}

\item{p}{(double scalar) Probability of obtaining an edge in the network.}

\item{seed}{(boolean scalar | int scalar) Should the sampling occur under seeding? Defaults to FALSE.}
}
\value{
A list containing the graph (i.e., interaction parameters) and the thresholds.
}
\description{
This function is used to generate random Ising model parameters for a random network architecture.
}
\details{
First, we sample a binary dataset N ^ 2, where N is the number of nodes in the network. The
probability of sampling a value of 1 is equal to the probability specified by the function
argument p, and the probability of sampling a 0 is 1 - p.

Next, we create a new symmetrical matrix from the sampled dataset and set its main diagonal
to 0. This represents the unweighted graph.

Next, we create an weighted graph from the unweighted graph. To do so we generate a matrix
with values from the normal distribution (M = 0, SD = 1) with the same number of columns
and rows and the unweighted matrix. We also make this matrix symmetrical (i.e., no need to
set the main diagonal to 0). Now, the weighted random graph is obtained by multiplying
the unweighted graph by the symmetrical random matrix.
}
