% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising_scale_free.R
\name{ising_scale_free}
\alias{ising_scale_free}
\title{Generate Ising parameters for a scale-free network architecture.}
\usage{
ising_scale_free()
}
\description{
This function is used to generate random Ising model parameters for a scale-free network architecture.
}
\details{
Not implemented.
}
