% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externals.R
\name{IsingFitEssential}
\alias{IsingFitEssential}
\title{.
Copyright: https://github.com/cvborkulo/IsingFit}
\usage{
IsingFitEssential(x, family = "binomial", AND = TRUE, gamma = 0.25,
  lowerbound.lambda = NA)
}
