% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{list_to_data_frames}
\alias{list_to_data_frames}
\title{.}
\usage{
list_to_data_frames(data)
}
\description{
Sorry for these messy loops. There is certainly a better way to do this: aperm(array(unlist(data), c(19, 1080, 118)), c(2, 1, 3))
}
