% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{list_to_data_frames_array}
\alias{list_to_data_frames_array}
\title{.}
\usage{
list_to_data_frames_array(data, cols, rows, layers, perm = c(2, 1, 3))
}
\description{
This should be the better alternative to the `list_to_data_frames()`.
}
