% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{netpaw}
\alias{netpaw}
\alias{netpaw-package}
\title{\code{netpaw}: A package for corss-sectional network simulations.}
\description{
The \code{netpaw} package provides three categories of functions...
TODO: Add documentation.
}
\section{\code{netpaw} functions}{

The \code{netpaw} functions...
}

\author{
\strong{Maintainer}: Mihai Constantin \email{constantin.mihai.alexandru@gmail.com}

}
