% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{netpower}
\alias{netpower}
\alias{netpower-package}
\title{netpower: A package for network simulations.}
\description{
The netpower package provides three categories of important functions:
TODO: fix the documenation for all functions.
}
\section{netpower functions}{

The netpower functions ...
}

\author{
\strong{Maintainer}: Mihai Constantin \email{constantin.mihai.alexandru@gmail.com}

}
