% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netPowerData}
\alias{netPowerData}
\title{netPower simulation data
The data is based on the simulation design and aggregated across 100 replications.}
\format{An object of class \code{matrix} with 1080 rows and 19 columns.}
\usage{
netPowerData
}
\description{
netPower simulation data
The data is based on the simulation design and aggregated across 100 replications.
}
\keyword{datasets}
