% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006_simulation.R
\name{run.design}
\alias{run.design}
\title{Run multiple design cell aka a `design` (not available to the end users).}
\usage{
run.design(design)
}
\description{
Run multiple design cell aka a `design` (not available to the end users).
}
