% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cells_replication.R
\name{run_cells_with_replication}
\alias{run_cells_with_replication}
\title{Runs the simulation for a selected number of design cells with replication.}
\usage{
run_cells_with_replication(cells, replications)
}
\arguments{
\item{cells}{(matrix) A matrix containing the specifications of the cells that will be ran.
Usually a subset of the result of \code{\link{build_desing}}.}

\item{replications}{(int scalar) The number of times to replicate the specified cells.}
}
\value{
A list containing, for each replication, the output of \code{\link{run_cell}}.
}
\description{
This function is a wrapper around \code{\link{run_cells}} and it applies the simulation procedure
for all specified specified, \strong{with replication}.
}
