% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{build_design}
\alias{build_design}
\title{Build the factorial design of the study.}
\usage{
build_design(participants, nodes, densities)
}
\arguments{
\item{participants}{(int vector) Varying number of participants, e.g., \code{c(10, 50, 100)}.}

\item{nodes}{(int vector) Varying number of nodes (i.e., variables), e.g., \code{c(10, 15, 30)}.}

\item{densities}{(double vector) Varying number of densities, e.g., \code{c(.2, .5, .8)}.}

\item{architecture}{(int vector) Network architectures, e.g., \code{c(random = 1, small_world = 2, scale_free = 3, empirical = 4)}.}
}
\value{
A matrix representing the factorial design of the study, where each row is a design cell.
}
\description{
This function builds the factorial design of the study based on the elements of the
vectors specified as arguments. Each argument specified (i.e., vector) is regarded
as a factor, and each value of the vectors is regarded as a level.

TODO: Add one more factor: network architecture (i.e., after asking Claudia).
}
