% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate.model}
\alias{estimate.model}
\title{Estimated the model based on the type of data provided.}
\usage{
estimate.model(data, memory.saver = TRUE)
}
\description{
Estimated the model based on the type of data provided.
}
