% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{netPower}
\alias{netPower}
\alias{netPower-package}
\alias{netPower-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The netPower package provides three categories of important functions:
foo, bar and baz.
}
\section{netPower functions}{

The netPower functions ...
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
