% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{netPower}
\alias{netPower}
\alias{netPower-package}
\alias{netPower-package}
\title{netPower: A package for network simulations.}
\description{
The netPower package provides three categories of important functions:
TODO: fix the documenation for all functions.
}
\section{netPower functions}{

The netPower functions ...
}

\author{
\strong{Maintainer}: Mihai Constantin \email{constantin.mihai.alexandru@gmail.com}

}
