% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cell.R
\name{run_cell}
\alias{run_cell}
\title{Run the simulation for a single configuration.}
\usage{
run_cell(participants, nodes, density)
}
\arguments{
\item{participants}{(int scalar) A number of participants, e.g., 100.}

\item{nodes}{(int scalar) A number of nodes (i.e., variables), e.g., 15.}

\item{density}{(double scalar) A value indicating the density, e.g., .3}
}
\value{
A list containing three sub-lists: \code{$config}, \code{$true}, and \code{$estimated}.
                     \code{$config} contains a vector with the levels specified for this cell (i.e., same order
                     as specified by the function definition). \code{$graph}) contains a matrix representing the
                     network structure. \code{$thresholds}) holds a vector representing the model thresholds.
}
\description{
This function performs the simulation procedure for a single combination of factor levels (e.g.,
100 participants x 15 nodes x .3 density). The simulation is ran without replication.
Check \code{run_cells_with_replication} for replication purposes.

TODO: Describe the steps of the simulation.
}
