% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cells.R
\name{run_cells}
\alias{run_cells}
\title{Run the simulation for the selected design cells.}
\usage{
run_cell(participants, nodes, density)
}
\arguments{
\item{cells}{(matrix) A matrix containing the specifications for the cells that will be ran.
Usually a subset of the result of \code{build_desing}.}
}
\value{
A list containing the output of \code{run_cell} for each cell specified.
}
\description{
This function is a wrapper around \code{run_cell} and it performs the simulation procedure
for all the design cells specified. The simulation is ran without replication. Check
\code{run_cells_with_replication} for replication purposes.

TODO: Improve the documentation: Link to the other functions.
}
